+++ storybook/storybook/stories/basics/1_concept/concept.stories.js
@@ -4,5 +4,4 @@ import { Concept } from './Concept';
 import { Styles } from './Styles';
 
 storiesOf('Basics|Начало работы', module)
-  .add('Введение', () => <Concept />)
-  .add('Как подключить стили', () => <Styles />);
+  .add('Введение', () => <Concept />);