+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.37.0] - 2020-12-21
 
+### Changed
+- Storybook: Changed library name to Korus-ui
+
+
 ## [0.36.0] - 2020-12-07
 
 ### Added